Issue #891 is fixed Update tenant_workspace_tags.py

+++ tenant_workspace/templatetags/tenant_workspace_tags.py
@@ -1030,59 +1030,59 @@ def render_question_type_051(workspace, module, node, question, answer):
     q1_picked = q1.content
 
     # Calculate annual totals.
-    total_yr1 = 0.0
-    total_yr2 = 0.0
-    total_yr3 = 0.0
+    total_yr1 = 0
+    total_yr2 = 0
+    total_yr3 = 0
     for item in q1_picked:
-        total_yr1 += float(item['var_5'])
-        total_yr2 += float(item['var_6'])
-        total_yr3 += float(item['var_7'])
+        total_yr1 += int(item['var_5'])
+        total_yr2 += int(item['var_6'])
+        total_yr3 += int(item['var_7'])
 
     autogen = {
         # --- YEAR 1 ---
         'yr_1': total_yr1,
-        'm_01': total_yr1 / 12.00,
-        'm_02': total_yr1 / 12.00,
-        'm_03': total_yr1 / 12.00,
-        'm_04': total_yr1 / 12.00,
-        'm_05': total_yr1 / 12.00,
-        'm_06': total_yr1 / 12.00,
-        'm_07': total_yr1 / 12.00,
-        'm_08': total_yr1 / 12.00,
-        'm_09': total_yr1 / 12.00,
-        'm_10': total_yr1 / 12.00,
-        'm_11': total_yr1 / 12.00,
-        'm_12': total_yr1 / 12.00,
+        'm_01': int(total_yr1 / 12),
+        'm_02': int(total_yr1 / 12),
+        'm_03': int(total_yr1 / 12),
+        'm_04': int(total_yr1 / 12),
+        'm_05': int(total_yr1 / 12),
+        'm_06': int(total_yr1 / 12),
+        'm_07': int(total_yr1 / 12),
+        'm_08': int(total_yr1 / 12),
+        'm_09': int(total_yr1 / 12),
+        'm_10': int(total_yr1 / 12),
+        'm_11': int(total_yr1 / 12),
+        'm_12': int(total_yr1 / 12),
 
         # --- YEAR 2 ---
         'yr_2': total_yr2,
-        'm_13': total_yr2 / 12.00,
-        'm_14': total_yr2 / 12.00,
-        'm_15': total_yr2 / 12.00,
-        'm_16': total_yr2 / 12.00,
-        'm_17': total_yr2 / 12.00,
-        'm_18': total_yr2 / 12.00,
-        'm_19': total_yr2 / 12.00,
-        'm_20': total_yr2 / 12.00,
-        'm_21': total_yr2 / 12.00,
-        'm_22': total_yr2 / 12.00,
-        'm_23': total_yr2 / 12.00,
-        'm_24': total_yr2 / 12.00,
+        'm_13': int(total_yr2 / 12),
+        'm_14': int(total_yr2 / 12),
+        'm_15': int(total_yr2 / 12),
+        'm_16': int(total_yr2 / 12),
+        'm_17': int(total_yr2 / 12),
+        'm_18': int(total_yr2 / 12),
+        'm_19': int(total_yr2 / 12),
+        'm_20': int(total_yr2 / 12),
+        'm_21': int(total_yr2 / 12),
+        'm_22': int(total_yr2 / 12),
+        'm_23': int(total_yr2 / 12),
+        'm_24': int(total_yr2 / 12),
 
         # --- YEAR 3 ---
         'yr_3': total_yr3,
-        'm_25': total_yr3 / 12.00,
-        'm_26': total_yr3 / 12.00,
-        'm_27': total_yr3 / 12.00,
-        'm_28': total_yr3 / 12.00,
-        'm_29': total_yr3 / 12.00,
-        'm_30': total_yr3 / 12.00,
-        'm_31': total_yr3 / 12.00,
-        'm_32': total_yr3 / 12.00,
-        'm_33': total_yr3 / 12.00,
-        'm_34': total_yr3 / 12.00,
-        'm_35': total_yr3 / 12.00,
-        'm_36': total_yr3 / 12.00
+        'm_25': int(total_yr3 / 12),
+        'm_26': int(total_yr3 / 12),
+        'm_27': int(total_yr3 / 12),
+        'm_28': int(total_yr3 / 12),
+        'm_29': int(total_yr3 / 12),
+        'm_30': int(total_yr3 / 12),
+        'm_31': int(total_yr3 / 12),
+        'm_32': int(total_yr3 / 12),
+        'm_33': int(total_yr3 / 12),
+        'm_34': int(total_yr3 / 12),
+        'm_35': int(total_yr3 / 12),
+        'm_36': int(total_yr3 / 12)
     }
 
     return {
